unix-Makefile.tmpl: make pkgconfig files honor shlib_variant
Presently on Linux, shlib_variant affects the filename, SONAME and symbol
versions of libcrypto.so and libssl.so. However, the pkgconfig files are
unaffected: they are still named libcrypto.pc and libssl.pc; they still specify
linking against libcrypto and libssl; etc.

If the developer specifies a namespaced openssl build via shlib_variant, it is
nonsensical to install non-namespaced pkgconfig files along with it. Therefore:

- Expose SHLIB_VARIANT appropriately in the Makefile
- Build libcrypto$(SHLIB_VARIANT).pc and libssl$(SHLIB_VARIANT).pc instead of
  libcrypto.so and libssl.so
- Inside the .pc files, take SHLIB_VARIANT into account in Name: (but not
  Description:) and Libs:

Signed-off-by: Rich Tollerton <rich.tollerton@ni.com>